refactor(backend): move middleware into own folder
