+++ packages/backend/src/middleware/lock/index.test.ts
@@ -1,8 +1,6 @@
-import Redis from 'ioredis'
 import pino from 'pino'
 import { v4 as uuid } from 'uuid'
-import { Config } from '../config/app'
-import { lockMiddleware, Lock, createRedisLock } from './lock'
+import { lockMiddleware, Lock } from '.'
 
 const createTestLock = (): Lock => {
   const map = new Map<string, string>()
@@ -131,116 +129,5 @@ describe('Lock Middleware', (): void => {
     })
     expect(lockAcquireSpy).toHaveBeenCalledTimes(3)
     expect(lockReleaseSpy).toHaveBeenCalledTimes(1)
-  })
-})
-
-describe('Redis Lock', (): void => {
-  const redis = new Redis(Config.redisUrl, {
-    tls: Config.redisTls,
-    stringNumbers: true
-  })
-
-  afterEach(async () => {
-    jest.useRealTimers()
-    await redis.flushall()
-  })
-
-  afterAll(async () => {
-    redis.disconnect()
-  })
-
-  describe('acquire', () => {
-    test('properly sets key', async () => {
-      const now = Date.now()
-      jest.useFakeTimers({ now })
-
-      const redisLockTtlMs = 10000
-      const keyPrefix = 'lock-prefix'
-      const redisLock = createRedisLock({
-        redisClient: redis,
-        keyTtlMs: redisLockTtlMs,
-        keyPrefix
-      })
-      const key = uuid()
-
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-      await expect(redis.get(`${keyPrefix}:${key}`)).resolves.toBe(
-        new Date(now + redisLockTtlMs).toISOString()
-      )
-    })
-
-    test('properly sets key with default prefix', async () => {
-      const now = Date.now()
-      jest.useFakeTimers({ now })
-
-      const redisLockTtlMs = 10000
-      const redisLock = createRedisLock({
-        redisClient: redis,
-        keyTtlMs: redisLockTtlMs
-      })
-      const key = uuid()
-
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-      await expect(redis.get(`lock:${key}`)).resolves.toBe(
-        new Date(now + redisLockTtlMs).toISOString()
-      )
-    })
-
-    test('returns false if lock already acquired', async () => {
-      const redisLockTtlMs = 10000
-      const redisLock = createRedisLock({
-        redisClient: redis,
-        keyTtlMs: redisLockTtlMs
-      })
-      const key = uuid()
-
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-      await expect(redisLock.acquire(key)).resolves.toBe(false)
-    })
-
-    test('lock expires properly', async () => {
-      const redisLockTtlMs = 100
-      const redisLock = createRedisLock({
-        redisClient: redis,
-        keyTtlMs: redisLockTtlMs
-      })
-      const key = uuid()
-
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-      await expect(redis.get(`lock:${key}`)).resolves.toBeDefined()
-      await new Promise((resolve) => setTimeout(resolve, redisLockTtlMs))
-      await expect(redis.get(`lock:${key}`)).resolves.toBeFalsy()
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-    })
-  })
-
-  describe('release', () => {
-    test('properly deletes key', async () => {
-      const redisLockTtlMs = 10000
-      const redisLock = createRedisLock({
-        redisClient: redis,
-        keyTtlMs: redisLockTtlMs
-      })
-      const key = uuid()
-
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-      await redisLock.release(key)
-
-      await expect(redis.get(`lock:${key}`)).resolves.toBeFalsy()
-    })
-
-    test('can re-acquire key after release', async () => {
-      const redisLockTtlMs = 10000
-      const redisLock = createRedisLock({
-        redisClient: redis,
-        keyTtlMs: redisLockTtlMs
-      })
-      const key = uuid()
-
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-      await redisLock.release(key)
-      await expect(redisLock.acquire(key)).resolves.toBe(true)
-      await expect(redis.get(`lock:${key}`)).resolves.toBeDefined()
-    })
   })
 })
